Implemented max evaluation of sum nodes
The eval function for every node now has a boolean  keyword argument max. If max==true, then the value of a sum node is not the weighted sum of its child values, but its maximum weighted child value.
This change was tested in test_eval_max in test/testnodes.jl.